mtree: Fix integer overflow on 32 bit systems
(cherry picked from commit b5aceb5887c03e22d34d21cd118232d38b300e00)

+++ libarchive/archive_read_support_format_mtree.c
@@ -337,7 +337,7 @@ next_line(struct archive_read *a,
 	 */
 	while (*nl == 0 && len == *avail && !quit) {
 		ssize_t diff = *ravail - *avail;
-		size_t nbytes_req = (*ravail+1023) & ~1023U;
+		size_t nbytes_req = ((size_t)*ravail + 1023) & ~1023U;
 		ssize_t tested;
 
 		/*
